chore: only affects 429
