+++ src/azure-cli/azure/cli/command_modules/network/aaz/latest/__init__.py
@@ -8,3 +8,4 @@
 # pylint: skip-file
 # flake8: noqa
 
+from ._clients import AAZNonRetryableClient
